move emails functionality to it's own sub-resource
